move more cam stuff out of sceneInfra
